pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "inkletter"
version = "0.1.0"
description = "Markdown to MJML/HTML converter and previewer"
authors = [
    { name = "Laurent Pauloin", email = "laurent@crococode.co" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click==8.2.1",
    "jinja2==3.1.6",
    "mistune==3.1.3",
    "mjml-python==1.3.5"
]

[project.scripts]
inkletter = "inkletter.cli:cli"

[tool.setuptools]
packages = ["inkletter"]